+++ server/order/go.mod
@@ -1,0 +1,9 @@
+module github.com/decred/dcrdex/server/order
+
+go 1.13
+
+require (
+	github.com/decred/dcrd/crypto/blake256 v1.0.0
+	github.com/decred/dcrdex/server/account v0.0.0-20190907171020-c48f5b8f4bbc
+	github.com/decred/slog v1.0.0
+)
